[Refactor] listStore 리펙터링 반영 #388
